refactor(dev): camelCase filenames
